+++ ui/src/app/modules/status/widgets/homebridge-status-widget/homebridge-status-widget.component.ts
@@ -86,4 +86,12 @@ export class HomebridgeStatusWidgetComponent implements OnInit {
       this.$toastr.error(err.message);
     }
   }
+
+  openUiSettings() {
+    this.$plugin.settings({
+      name: 'homebridge-config-ui-x',
+      settingsSchema: true,
+      links: {},
+    });
+  }
 }